requirements.txt
Django>=5.0,<6.0
gunicorn
whitenoise
django-apscheduler
django-solo
Pillow
